[] Add the CreateLoadComplexTable example program to the nosql-examples/examples-nosql-java-sdk gitbub repo
First set of changes recommended by Dario before merging PullRequest

M  README.md

Added requirement to the Java Version section that tells the user to either set the PATH or JAVA_HOME environment variable to point to the Java installation so the build can find java, javac, and javadoc in the installation. Also, two minor typo fixes.

M  src/main/java/nosql/cloud/table/CreateLoadComplexTable.java

Removed the REGIONS_MAP and change to calling the Region.fromRegionId() method.

M  CreateLoadComplexTable/src/test/java/nosql/cloud/table/ComplexTableTestBase.java

Fixed an issue discovered by Dario in which using the "java.io.tmpdir" as the base for the TEST_DIR causes problems in some environments with different OS's. For example, on MacOS, java.io.tmpdir appends a slash ('/') to the returned value, whereas on Oracle linux, it does not include the trailing slash. To address this issue, changed the base directory for TEST_DIR to be the value of the current working directory ("user.dir"); which does not have the same sort of inconsistencies on the different OS's.